third_party/blink/web_tests/http/tests/eventsource/workers/eventsource-cors-no-server-expected.txt

[Worker] Test that EventSource tries to reconnect if there's no server response when making cross-origin requests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-cors-no-server.js
PASS [Worker] es = new EventSource("http://127.0.0.1:12345/event-stream") did not throw exception.
PASS [Worker] got error event and readyState is CONNECTING
PASS [Worker] es = new EventSource("http://localhost:54321/event-stream") did not throw exception.
PASS [Worker] got error event and readyState is CONNECTING
PASS successfullyParsed is true

TEST COMPLETE
